Move CompositorMsg to script_traits.
